Consolidate all library requires into a single pantry.rb
I've no reached the point where requires are getting unwieldy and
leading to a ton of duplication and possibility of getting the path
wrong. There's also no need in this day and age to force a user of a
library or application to require more than one file to get the whole
system.

Not doing any auto-loading. This will let me keep very clear limits of
what is included in the system and what isn't.