Refactor how the system sends messages to intended recipients
Passing around the Filter objects was getting cumbersome and confusing.
In the end, what has to get sent down to ZMQ is a single stream or
identity string. Realizing that what I want is really no different from
Email, I now follow SMTP's pattern and changed Message to have #to and #from,
which contains those values, and updated Server / Client / etc to work with those
fields almost explicitly.

Updated ListClients and the CLI setup of this command to move the
filters used to choose which clients to return into the message body,
letting me remove the #filters metadata from Messages

Disabled most of the CLI acceptance tests as I'm still figuring out how
to manage the data return from the clients/server appropriately.