Convert Pantry::Client and Pantry::Server to be proper Actors
Probably just an oversight, these two objects definitely need to be
Celluloid actors as I plan on using Timers in them as well as
Supervisors to keep them alive when I get the run scripts in place.

This change then broke the entire use of #on for custom event handlers
so I had to find a replacement for them. Annoying but I figured they
would go away soon enough as they were just to help me test early on. As
a replacement I'm now fully using the Commands system and have
implemented a simple Echo command that returns the message untouched
back to the sender.

Unfortunately as Client and Server are now Celluloid Actors, all unit
tests need to be sure that Celluloid is started up. I don't think
there's really an issue here speed wise but will keep an eye on things.