Remove ExecuteShell command
I'm not sure that a blanket shell execution command is a good thing to
have in this project. For now I'm killing it and will revisit if we need
to be running random calls against servers.

Replaced most of the tests to instead use the Echo command. However
given this Echo command was added to client and server had problems with
targetting server vs client. The solution, a server-only echo in the
tests, feels hacky. I may revisit the global-ness of these command
lists.