Send client list response back to requester before forwarding messages
The CLI process, which should probably be redesigned, is two-fold for
requests meant for Clients. The CLI after sending the request waits for
the Server to respond with a list of Clients that will be responding.
The CLI then waits for those Clients to respond back with their
information. On some systems this Server response *after* forwarding was
also ensuring that Clients responded before the Server, and due to
Message UUID response mapping, the CLI got the first Client response as
the Server response and all hell broke loose.

I'll be working to think up and try new ways of handling this
asyncronicity.