Add our own check that the incoming message is meant for the Client
I didn't fully understand just how naiive zmq's pub/sub topic matching
is. Basically if *any* subscribe matches the topic, the message goes
through, so if we want a message to only got to, say, pantry.test.app,
we have to check twice because "pantry" and "pantry.test" matches. We
can't, however, just subscribe to "pantry.test.app" because then
messages to "pantry" or "pantry.test" do *not* match.

There's discussion of making this matching smarter, but this does not
seem to be something that will end up in ZMQ core any time soon. This
check also helps make the core less ZMQ dependent.

Along this, went ahead and made the default Message#to value to the
catch-all stream "" to ensure we don't run into nil errors.