Re-think how CLI and commands are structured
In starting to set up the Chef Cookbook Upload command I realized the
current setup of how CLI built and sent commands was too naiive for this
new command. After some thinking and playing I've decided to move more
logic into Command objects and update the CLI to use this new Command
API. There are still some warts, this is definitely a "get it working
again" half-way step and more rework is on the way. This commit does the
following:

* Commands can now run code and control Message creation via #prepare_message.
* Commands are also in charge of how to handle the response Messages. By
  default this is to simply return the response Message.
* Moved Multi-response messages out and added it to the Echo Command as
  an example of how to change the response handling. This will be
  updated.
* Got rid of the lambda hack for ListClients as that got moved into
  #prepare_message! Also no longer care if a message is server-only. The
  assumption now is that a command is only client or only server as
  that's all I've made so far.