Update WaitList to support persistent message receivers
As I move into file uploads I need a system that can handle sending
multiple sequential messages to a single object. This now exists for the
WaitList under persistent management.

There may be a race condition between receiving a message and checking
if the worker is finished. I can't get the tests to fail so I'll keep an
	eye on this and assume for now it's working.