Re-attack how Commands are managed by the CLI
So I couldn't find a good way to ensure Commands always all #finish on
their listener by default. This led me to realize that the #handle_response
idea was putting too much responsibility on the command. Instead of
using the wait_list request cycle here, I could just send and recieve
messages, as it's assumed these messages are supposed to end up at this
client. This then massively cleared up the lock issue with commands not
calling #finished.

Now, the default behavior for #receive_response is to send the response
message to the current listener and mark the command as finished. I also
made the listener configurable in CLI so that it can be used in tests,
making it possible to test what the output is going to be from these
commands.