Delegate Client information to another object
When I wrote this I didn't realize .new actually started a new Actor in
the system, so using full Pantry::Clients when registering a client on
the server was actually creating new actors every call. As there's no
need for any of that, but there is need for the client's identifying
information in the ClientRegistery, I decided to pull this info into a
helper object, delegate to that object, and use it when registering.