Rebuild file service to be single actors for send and receive
After fighting with my "start an actor on demand for file processing"
design for a while I've decided to change this up a bit. Now, the
FileService is itself an actor that communicates on it's own set of
sockets. It runs a Sender and a Receiver that each manage their
perspective jobs for whatever, and however many, files that are being
processed.

This commit doesn't fix up the rest of the library though. I wanted a
commit that was just the underlying changes to file serving. Next will
be cleaning up the use of this new API.