Update file progres objects to include finish checks
This logic was used in Progres Listeners previously but it made a lot
more sense to have the sender/receiver give back an object that will
contain this information. The previous design of passing around a
listener object felt very heavy handed and was requiring a lot of
objects to know about listeners, which I didn't like at all.