Change storage of Curve keys to live in a single file
Figuring out how to work with Curve keys on the CLI prompted this change
and I think it's better as a whole anyway. Now, all keys required to run
a client are stored in a single file, client_keys.yml. For now this will
require editing the generated file after the fact but that won't last as
full client authentication will ensure that the file will be written out
complete when asking for new client credentials.