Change data_dir default to nil
Instead of setting a default in the code, going to instead set the
default in the dist/ client and server config files. This in-code
default has been a problem with testing CLI and the .pantry (which I
want to also be a data_dir for the CLI for Curve keys purposes). By
using a nil default here I can better have the CLI choose to use a
pre-set data dir or use a .pantry dir.