Add UI output helper class
This will be used to support STDOUT in commands and elsewhere as a way
to notify the user what's going on when you don't want to use
Pantry.logger. I've chosen to use Highline as the handler, which will
take care of a number of possible situations I might need.