Replace ProgressListeners with Pantry::UI and Command#finished
I never was happy with the progress listener setup, and recently
realized that I didn't need it at all, but could instead let the Command
know when it was finished. With this change, and with the new Pantry::UI
setup I've just added, the crappy listener stuff can finally go.