Designate which message parts are actually JSON
So this is a pseudo hack-ish way of type designation but it's simple and
encapsulated entirely in this object. I ran into a situation where I was
uploading a chef data bag, which is JSON. When the content reached the
server, it was picked up as JSON, parsed into a Hash, then the inspect
of that Hash was written out to the final file.

I'll be playing with this to make sure the forced UTF8 encoding doesn't
cause any serious problems. If so I'll have to drop this back to some
random ASCII designation.