Store the first connecting client's public key
A new server will have an empty list of known client keys so no-one can
connect, thus making the system useless right off the bat and quite
difficult to fix. Thus to facilitate this, the first client to connect
ot a server is immediately marked as an authenticated client. This
client will still need to be given the server's public key but can
autogenerate it's own keys, which I'll get working next.